+++ CHANGELOG.md
@@ -9,6 +9,7 @@
  * Added better logging to all test method output to make it clear what property is causing a test to fail
  * Added support for NetBIOS domain names resolution to SPUserProfileServiceApp
  * Removed chocolatey from the AppVeyor build process in favour of the PowerShell Gallery build of Pester
+ * Added resource MSFT_SPTrustedIdentityTokenIssuer
 
 ### 1.1
 
@@ -122,5 +123,3 @@
 
 * Initial public release of xSharePoint
 
-### Unreleased
-- Added resource MSFT_SPTrustedIdentityTokenIssuer